Ignore previous frame in mp3_mute_frame
mp3_mute_frame manipulates the aux_size of the previous frame, but mp3_unmute_frame doesn't.
This leads to an inconsistency between aux_size in the encoder and decoder. In most cases this doesn't seem to be a problem.
However, in some rare cases, it leads to packMP3 crashing while trying to uncompress the .pmp file.